2025/March13-Intro/requirements.txt
flask==3.1.0
google-generativeai==0.8.4
